Fixes for GDB+Insight (6.7.1):
Added 'gdb_grab', which will correctly search for 6.7.1 (ftp://ftp.gnu.org/gnu/gdb/), as there exists only 6.7.1a. Changed a few 'gdb_ver' to 'gdb_grab' instead (starts at line 218). This ensures it downloads the correct version that is listed in the FTP, and when it mkdirs GDB, it correctly creates the directory as './gdb-6.7.1' and NOT './gdb-6.7.1a', as it won't work with the pre-existing 'gdb_ver' define. The exact same method was done for Insight (creating 'insige_grab') and changing a few defines in the INSIGHT BUILD script. Once these are defined, the builds link correctly with no issues, and lessens the work the end-user might have to do to get gdb/insight pulled and downloaded correctly.

Tested and verified with the latest git pull of KallistiOS under Cygwin x86/Windows 10, which was built successfully from the instructions provided at this link:
http://dcemulation.org/phpBB/viewtopic.php?p=1042772#p1042772